app_sensors: add support for Thingy91x
Add conditionals to handle sensor variations between the Thingy91 and the
Thingy91x.

Known issues: the iaq/co2/voc readings for the BME68x on the Thingy91 don't
appear to ever change. However, this is the same behavior observed for the
factory firmware of the Thingy91x so it is being added for now.

Signed-off-by: Mike Szczys <mike@golioth.io>