thing91x: configure the BME688 for more frequent readings
The BME68x driver was configured to only allow the sensor to perform
readings every 5 minutes. This PR changes that setting to take readings
every 3 seconds.

This is a configuration for the sensor hardware itself. It is separate from
the Golioth LOOP_DELAY_S that configures how often readings are sent to
Golioth.

Signed-off-by: Mike Szczys <mike@golioth.io>